Make more modular and improve readability
