+++ server.js
@@ -21,7 +21,8 @@ app.use(session({
     resave: false,
     saveUninitialized: true,
     cookie: { 
-        secure: process.env.SESSION_SECURE === 'true',
+        // secure: process.env.SESSION_SECURE === 'true',
+        secure: false,
         httpOnly: true,
         proxy: true,
         maxAge: 24 * 60 * 60 * 1000
